+++ ownership-chain/runtime/laos/src/configs/evm.rs
@@ -1,4 +1,4 @@
-use super::fee::ToAuthor;
+use super::types::ToAuthor;
 use crate::{
 	precompiles::FrontierPrecompiles, AccountId, Aura, Balances, BaseFee, EVMChainId, Runtime,
 	RuntimeEvent, Timestamp, Weight, U256, WEIGHT_REF_TIME_PER_SECOND,
