+++ ownership-chain/runtime/laos/src/configs/mod.rs
@@ -3,11 +3,11 @@ use ownership_parachain_primitives::Balance;
 
 mod block_rewards_handler;
 mod evm;
-pub mod fee;
 mod multisig;
 pub mod parachain_staking;
 mod proxy;
 mod transaction_payment;
+pub mod types;
 mod utility;
 
 // Define storage fees as constants for clarity and reuse
